+++ mods/BrewLAN/modules/BrewUI/strategicicons/mod_info.lua
@@ -1,0 +1,14 @@
+name = "BrewLAN UI: Strategic Icon Overhaul"
+uid = "Z529ea71-93ef-41a6-b552-STRATICON00010"
+version = 1
+copyright = "© 2015–2021 Sean Wheeldon"
+description = "— Requires SCD — Applied sim-side —\nReassigns strategic icons based on physical properties of the units."
+author = "Balthazar"
+icon = "/mods/BrewLAN/modules/BrewUI/strategicicons/mod_icon.png"
+selectable = true
+enabled = true
+exclusive = false
+ui_only = false
+conflicts = {  }
+before = {  }
+after = { "25D57D85-7D84-27HT-A501-BR3WL4N000089" }